Revert "GC migration fases A/B/0/1: per-message router + Jint JS engine + pro…"
