--- gc-js/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "skynet-gc-js",
-  "version": "0.1.0",
-  "private": true,
-  "description": "Authoring project (dev/CI only) for the SKYNET Game Coordinator JS layer. The server runs the committed bundle under Jint; Node is never required at runtime.",
-  "type": "module",
-  "scripts": {
-    "typecheck": "tsc --noEmit",
-    "proto:gen": "pbjs -t json -o \"../SKYNET server/GC/570/js/proto-descriptor.json\" ../proto/*.proto",
-    "build:bundle": "esbuild src/index.ts --bundle --format=iife --platform=browser --outfile=\"../SKYNET server/GC/570/js/dist/gc.js\"",
-    "build": "node -e \"require('node:fs').mkdirSync('../SKYNET server/GC/570/js/dist', { recursive: true })\" && npm run proto:gen && npm run build:bundle",
-    "gen:inventory": "node tools/gen-inventory.mjs",
-    "check:sync": "node tools/check-bundle-sync.mjs",
-    "check:gates": "node tools/check-gates.mjs"
-  },
-  "devDependencies": {
-    "esbuild": "0.25.5",
-    "long": "5.3.2",
-    "protobufjs": "7.5.3",
-    "protobufjs-cli": "1.1.3",
-    "typescript": "5.8.3"
-  }
-}
