--- gc-js/quarantine.json
@@ -1,205 +1,0 @@
-{
-  "$comment": "Generated by gc-js/tools/gen-inventory.mjs. Names of C#/Lua response-builder and fixture-replay members that migrated message routes must never call. Enforced by gc-js/tools/check-gates.mjs.",
-  "forbidden": [
-    "AccountGuildPersonaInfoBatchResponse",
-    "AccountGuildPersonaInfoResponse",
-    "AvailablePrivateCoachingSessionsResponse",
-    "AvailablePrivateCoachingSessionsSummaryResponse",
-    "BuildAccountGuildPersonaInfoBatchResponse",
-    "BuildAccountGuildPersonaInfoResponse",
-    "BuildAccountGuildsPersonaInfo",
-    "BuildAvailablePrivateCoachingSessionsResponse",
-    "BuildAvailablePrivateCoachingSessionsSummaryResponse",
-    "BuildBatchPlayerResourcesResponse",
-    "BuildBoolField",
-    "BuildChatRegionsEnabled",
-    "BuildClientWelcome",
-    "BuildConnectString",
-    "BuildConnectionStatus",
-    "BuildConnectionStatusHaveSession",
-    "BuildDotaGameAccountClient",
-    "BuildDotaGameAccountPlus",
-    "BuildDotaItemInstanceId",
-    "BuildDotaStatsGameMatchSignOutResponse",
-    "BuildDotaStatsMatch",
-    "BuildDotaStatsMatchPlayer",
-    "BuildDotaStatsPlayerStatsResponse",
-    "BuildDotaStatsProfileCard",
-    "BuildDotaStatsProfileResponse",
-    "BuildDotaStatsShowcase",
-    "BuildDotaStatsShowcaseStatItem",
-    "BuildDotaStatsSocialComment",
-    "BuildEconGameAccountClient",
-    "BuildEconInventoryPositions",
-    "BuildEconItem",
-    "BuildEconItemEquipped",
-    "BuildEconItemIdOnly",
-    "BuildEconItemsForUser",
-    "BuildEconMultipleObjects",
-    "BuildEconSingleObject",
-    "BuildEconSoCacheSubscribed",
-    "BuildEquipItemsResponse",
-    "BuildEventPointsResponse",
-    "BuildExtraMessage",
-    "BuildFindTopSourceTvGamesResponse",
-    "BuildGameMatchSignOutResponse",
-    "BuildGameServerSoCacheSubscribed",
-    "BuildGameServerWelcome",
-    "BuildGameSoCacheSubscribed",
-    "BuildGuildMembershipResponse",
-    "BuildHeroRecentAccomplishments",
-    "BuildHeroStatsHistoryResponse",
-    "BuildInvitationCreated",
-    "BuildItemAge",
-    "BuildItemAges",
-    "BuildJoinResponse",
-    "BuildLobbyBroadcastObject",
-    "BuildLobbyMember",
-    "BuildLobbyMemberSummaryObject",
-    "BuildLobbyMultipleObjects",
-    "BuildLobbyObject",
-    "BuildLobbyPersonaObject",
-    "BuildLobbySelectionPriorityRule",
-    "BuildLobbySingleObject",
-    "BuildLobbySoCacheSubscribed",
-    "BuildLobbySoCacheUnsubscribed",
-    "BuildLocation",
-    "BuildMatchGroupsVersion",
-    "BuildMatchSignedOut",
-    "BuildMatchmakingStatsResponse",
-    "BuildNotificationsResponse",
-    "BuildOwnerSoid",
-    "BuildPartyInviteObject",
-    "BuildPartyInviteSubscribed",
-    "BuildPartyInviteUnsubscribed",
-    "BuildPartyMemberObject",
-    "BuildPartyMultipleObjects",
-    "BuildPartyObject",
-    "BuildPartySingleObject",
-    "BuildPartySingleObjectDestroyed",
-    "BuildPartySoCacheSubscribed",
-    "BuildPartySoCacheUnsubscribed",
-    "BuildPlayerMatchRecord",
-    "BuildPlayerRecentAccomplishments",
-    "BuildPlayerRecentCommends",
-    "BuildPlayerRecentMatchInfo",
-    "BuildPlayerRecentMatchOutcomes",
-    "BuildProfileCardForAccount",
-    "BuildProfileCardHeroSlot",
-    "BuildProfileCardResponse",
-    "BuildProfileCardStatSlot",
-    "BuildProfileResponse",
-    "BuildProfileTicketsResponse",
-    "BuildReadyCheckObject",
-    "BuildRealtimeStatsStartStop",
-    "BuildRecentAccomplishments",
-    "BuildRecentAccomplishmentsResponse",
-    "BuildRecentCommends",
-    "BuildRecentMatchInfo",
-    "BuildRecentOutcomes",
-    "BuildRecentRecord",
-    "BuildResult",
-    "BuildRoleHandicap",
-    "BuildSoCacheSubscribed",
-    "BuildSoCacheSubscribedForOwner",
-    "BuildSoCacheSubscribedUpToDate",
-    "BuildSoCacheSubscriptionCheck",
-    "BuildSoCacheUnsubscribed",
-    "BuildStableChatChannelId",
-    "BuildStoreSalesDataResponse",
-    "BuildSubscribedType",
-    "BuildTimedAverages",
-    "BuildTimedStdDevs",
-    "BuildTopCustomGamesList",
-    "BuildUnlockedEmoticonMask",
-    "BuildWeekendTourneyDivision",
-    "BuildWeekendTourneySchedule",
-    "ClientWelcomeResponse",
-    "Concat",
-    "ConnectionStatus",
-    "ConnectionStatusHaveSession",
-    "DotaEmoticonDataResponse",
-    "DotaEquipItemsResponse",
-    "DotaHeroStickersResponse",
-    "DotaJoinChatChannelResponse",
-    "DotaPartyInviteResponse",
-    "DotaPeriodicResourceResponse",
-    "DotaQuestProgressResponse",
-    "DotaSetItemStyleResponse",
-    "DotaStatsAllHeroOrderResponse",
-    "DotaStatsAllHeroProgressResponse",
-    "DotaStatsBatchPlayerResourcesResponse",
-    "DotaStatsClientHeroRecentAccomplishmentsResponse",
-    "DotaStatsClientRecentAccomplishmentsResponse",
-    "DotaStatsConductScorecardResponse",
-    "DotaStatsEventPointsResponse",
-    "DotaStatsGameMatchSignOutResponse",
-    "DotaStatsHasMvpVoteResponse",
-    "DotaStatsHeroGlobalDataResponse",
-    "DotaStatsHeroStandingsResponse",
-    "DotaStatsHeroStatsHistoryResponse",
-    "DotaStatsHeroTimedStatsResponse",
-    "DotaStatsLookupAccountNameResponse",
-    "DotaStatsMatchDetailsResponse",
-    "DotaStatsMatchVoteResponse",
-    "DotaStatsMvpVoteTimeoutResponse",
-    "DotaStatsPlayerMatchHistoryResponse",
-    "DotaStatsPlayerStatsResponse",
-    "DotaStatsProfileCardResponse",
-    "DotaStatsProfileResponse",
-    "DotaStatsProfileTicketsResponse",
-    "DotaStatsProfileUpdateResponse",
-    "DotaStatsRankResponse",
-    "DotaStatsServerRecentAccomplishmentsResponse",
-    "DotaStatsSetMatchHistoryAccessResponse",
-    "DotaStatsShowcaseGetUserDataResponse",
-    "DotaStatsSocialFeedCommentsResponse",
-    "DotaStatsSocialFeedPostCommentResponse",
-    "DotaStatsSocialFeedResponse",
-    "DotaStatsSocialMatchDetailsResponse",
-    "DotaStatsSocialMatchPostCommentResponse",
-    "DotaStatsSubmitLobbyMvpVoteResponse",
-    "DotaStatsSubmitPlayerReportResponse",
-    "DotaStatsTeammateStatsResponse",
-    "DotaStatsTrophyListResponse",
-    "DotaStatsVoteForMvpResponse",
-    "DotaTeamsInfoResponse",
-    "EconSoCacheSubscribed",
-    "Encode",
-    "EventPointsResponse",
-    "FieldBool",
-    "FieldBytes",
-    "FieldCount",
-    "FieldFixed32",
-    "FieldFixed64",
-    "FieldFixed64String",
-    "FieldString",
-    "FieldVarint",
-    "FieldVarintString",
-    "Fixture",
-    "GameSoCacheSubscribed",
-    "GuildMembershipResponse",
-    "MatchmakingStatsResponse",
-    "MessageWithTargetJob",
-    "MessageWithTargetJobString",
-    "NotificationsResponse",
-    "ProfileCardResponse",
-    "ProfileCardUpdate",
-    "ProfileResponse",
-    "ProfileTicketsResponse",
-    "Proto",
-    "Raw",
-    "Reply",
-    "ReplyEmpty",
-    "Result",
-    "StoreSalesDataResponse",
-    "WriteBytesField",
-    "WriteFixed32Field",
-    "WriteFixed64Field",
-    "WriteFloatField",
-    "WriteStringField",
-    "WriteVarint",
-    "WriteVarintField",
-    "observed_fixture"
-  ]
-}
