--- proto/base_gcmessages.proto
@@ -1,35 +1,0 @@
-// Curated subset of the community econ ("base") GC protobufs for app 570.
-// Field numbers validated against nethook fixtures by the spike suite.
-syntax = "proto2";
-
-message CSOEconItemAttribute {
-  optional uint32 def_index = 1;
-  optional uint32 value = 2;
-  optional bytes value_bytes = 3;
-}
-
-message CSOEconItemEquipped {
-  optional uint32 new_class = 1;
-  optional uint32 new_slot = 2;
-}
-
-message CSOEconItem {
-  optional uint64 id = 1;
-  optional uint32 account_id = 2;
-  optional uint32 inventory = 3;
-  optional uint32 def_index = 4;
-  optional uint32 quantity = 5;
-  optional uint32 level = 6;
-  optional uint32 quality = 7;
-  optional uint32 flags = 8 [default = 0];
-  optional uint32 origin = 9;
-  repeated CSOEconItemAttribute attribute = 12;
-  optional CSOEconItem interior_item = 13;
-  optional uint32 style = 15;
-  optional uint64 original_id = 16 [default = 0];
-  repeated CSOEconItemEquipped equipped_state = 18;
-}
-
-message CMsgDOTAGenericResult {
-  optional uint32 eresult = 1 [default = 2];
-}
