--- proto/gcsdk_gcmessages.proto
@@ -1,103 +1,0 @@
-// Curated subset of the community GC SDK protobufs (app 570).
-// Source of truth for the GC JS layer. Field numbers are validated by
-// round-tripping the nethook fixtures (spikes/gc-jint-protobuf); do not
-// change a field without re-running that suite.
-syntax = "proto2";
-
-enum GCConnectionStatus {
-  GCConnectionStatus_HAVE_SESSION = 0;
-  GCConnectionStatus_GC_GOING_DOWN = 1;
-  GCConnectionStatus_NO_SESSION = 2;
-  GCConnectionStatus_NO_SESSION_IN_LOGON_QUEUE = 3;
-  GCConnectionStatus_NO_STEAM = 4;
-}
-
-message CMsgSOIDOwner {
-  optional uint32 type = 1;
-  optional uint64 id = 2;
-}
-
-message CMsgSOSingleObject {
-  optional int32 type_id = 2;
-  optional bytes object_data = 3;
-  optional fixed64 version = 4;
-  optional CMsgSOIDOwner owner_soid = 5;
-}
-
-message CMsgSOMultipleObjects {
-  message SingleObject {
-    optional int32 type_id = 1;
-    optional bytes object_data = 2;
-  }
-
-  repeated SingleObject objects_modified = 2;
-  optional fixed64 version = 3;
-  repeated SingleObject objects_added = 4;
-  repeated SingleObject objects_removed = 5;
-  optional CMsgSOIDOwner owner_soid = 6;
-}
-
-message CMsgSOCacheSubscribed {
-  message SubscribedType {
-    optional int32 type_id = 1;
-    repeated bytes object_data = 2;
-  }
-
-  repeated SubscribedType objects = 2;
-  optional fixed64 version = 3;
-  optional CMsgSOIDOwner owner_soid = 4;
-}
-
-message CMsgSOCacheUnsubscribed {
-  optional CMsgSOIDOwner owner_soid = 1;
-}
-
-message CMsgSOCacheSubscribedUpToDate {
-  optional fixed64 version = 1;
-  optional CMsgSOIDOwner owner_soid = 2;
-}
-
-message CMsgSOCacheHaveVersion {
-  optional CMsgSOIDOwner soid = 1;
-  optional fixed64 version = 2;
-  optional int32 service_id = 3;
-  optional uint64 cached_file_version = 4;
-}
-
-message CMsgClientHello {
-  optional uint32 version = 1;
-  repeated CMsgSOCacheHaveVersion socache_have_versions = 2;
-  optional uint32 client_session_need = 3;
-  optional uint32 client_launcher = 4;
-  optional string secret_key = 5;
-  optional uint32 client_language = 6;
-}
-
-message CMsgClientWelcome {
-  message Location {
-    optional float latitude = 1;
-    optional float longitude = 2;
-    optional string country = 3;
-  }
-
-  optional uint32 version = 1;
-  optional bytes game_data = 2;
-  repeated CMsgSOCacheSubscribed outofdate_subscribed_caches = 3;
-  repeated CMsgSOCacheHaveVersion uptodate_subscribed_caches = 4;
-  optional Location location = 5;
-  optional bytes game_data2 = 6;
-  optional uint32 rtime32_gc_welcome_timestamp = 7;
-  optional uint32 currency = 8;
-  optional uint32 balance = 9;
-  optional string balance_url = 10;
-  optional string txn_country_code = 11;
-}
-
-message CMsgConnectionStatus {
-  optional GCConnectionStatus status = 1;
-  optional uint32 client_session_need = 2;
-  optional int32 queue_position = 3;
-  optional int32 queue_size = 4;
-  optional int32 wait_seconds = 5;
-  optional int32 estimated_wait_seconds_remaining = 6;
-}
