--- proto/spike_test.proto
@@ -1,28 +1,0 @@
-// Synthetic messages used only by spikes/gc-jint-protobuf to exercise the
-// difficult protobuf shapes under Jint (repeated fixed64, packed repeated,
-// bytes, unknown enum values, nested messages, sint64 zigzag).
-// Not part of the GC wire contract.
-syntax = "proto2";
-
-enum SpikeEnum {
-  SPIKE_ZERO = 0;
-  SPIKE_ONE = 1;
-}
-
-message SpikeNested {
-  optional fixed64 id = 1;
-  optional string name = 2;
-}
-
-message SpikeKitchenSink {
-  optional uint64 u64 = 1;
-  optional fixed64 f64 = 2;
-  repeated fixed64 f64_list = 3;
-  repeated uint64 packed_u64 = 4 [packed = true];
-  optional bytes blob = 5;
-  optional SpikeEnum enum_val = 6;
-  optional SpikeNested nested = 7;
-  repeated SpikeNested nested_list = 8;
-  optional sint64 s64 = 9;
-  optional string text = 10;
-}
